// frontend/src/api/generated/schemas/bookingStatusEnum.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * ParkMe API
 * Universal Parking Management Platform - Production API
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `PENDING` - Pending Payment
* `CONFIRMED` - Confirmed
* `ACTIVE` - Currently Parked
* `COMPLETED` - Completed
* `CANCELLED` - Cancelled
* `NO_SHOW` - No Show
 */
export type BookingStatusEnum = typeof BookingStatusEnum[keyof typeof BookingStatusEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const BookingStatusEnum = {
  PENDING: 'PENDING',
  CONFIRMED: 'CONFIRMED',
  ACTIVE: 'ACTIVE',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED',
  NO_SHOW: 'NO_SHOW',
} as const;