// frontend/src/api/generated/schemas/bookingTypeEnum.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * ParkMe API
 * Universal Parking Management Platform - Production API
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `RESERVATION` - Advance Reservation
* `DRIVE_IN` - Drive-In (Walk-up)
 */
export type BookingTypeEnum = typeof BookingTypeEnum[keyof typeof BookingTypeEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const BookingTypeEnum = {
  RESERVATION: 'RESERVATION',
  DRIVE_IN: 'DRIVE_IN',
} as const;