// frontend/src/api/generated/schemas/bookingRequest.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * ParkMe API
 * Universal Parking Management Platform - Production API
 * OpenAPI spec version: 1.0.0
 */
import type { BookingTypeEnum } from './bookingTypeEnum';
import type { BookingStatusEnum } from './bookingStatusEnum';

/**
 * Full booking details
 */
export interface BookingRequest {
  /** @nullable */
  actual_exit?: Date | null;
  booking_type?: BookingTypeEnum;
  entry_time: Date;
  /**
   * @nullable
   * @pattern ^-?\d{0,8}(?:\.\d{0,2})?$
   */
  estimated_price?: string | null;
  /** @nullable */
  expected_exit?: Date | null;
  facility: string;
  /**
   * @minimum 0
   * @maximum 2147483647
   */
  grace_period_minutes?: number;
  metadata?: unknown;
  /** @nullable */
  payment?: string | null;
  spot: string;
  status?: BookingStatusEnum;
  user: string;
  vehicle: string;
}