// frontend/src/api/generated/schemas/vehicleTypeEnum.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * ParkMe API
 * Universal Parking Management Platform - Production API
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `MOTORCYCLE` - Motorcycle
* `CAR` - Car
* `SUV` - SUV
* `TRUCK` - Truck
* `ELECTRIC_CAR` - Electric Car
* `ELECTRIC_SUV` - Electric SUV
* `BICYCLE` - Bicycle
* `VAN` - Van
* `BUS` - Bus
 */
export type VehicleTypeEnum = typeof VehicleTypeEnum[keyof typeof VehicleTypeEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const VehicleTypeEnum = {
  MOTORCYCLE: 'MOTORCYCLE',
  CAR: 'CAR',
  SUV: 'SUV',
  TRUCK: 'TRUCK',
  ELECTRIC_CAR: 'ELECTRIC_CAR',
  ELECTRIC_SUV: 'ELECTRIC_SUV',
  BICYCLE: 'BICYCLE',
  VAN: 'VAN',
  BUS: 'BUS',
} as const;